+++ assets/auth_hooks.py
@@ -18,7 +18,7 @@ class AssetsMenuItem(MenuItemHook):
         super().__init__(
             f"{app_settings.ASSETS_APP_NAME}",
             "fas fa-building-user fa-fw",
-            "assets:index",
+            "assets:location",
             navactive=["assets:"],
         )